Break out JSON API calls to separate files
